+++ IBS.Models/MSAP/Collection.cs
@@ -102,6 +102,8 @@ namespace IBS.Models.MSAP
 
         #endregion
 
+        [NotMapped]
+        public bool IsMonthClosed { get; set; }
     }
 }